Features for the future, bugs of the past
Added options to set dis(appear) animation type (slide/fade) and the
directions it should slide to (up, right, down or left). An extra option
is added to toggle resize animation.

Callbacks are now defined as a data attribute in the HTML element and
stored in an array in the options.

Fixed a bug where disabling the close button would disable the
"recenter" function causing the modal to be at upper left corner all the
time.

Fixed a bug where clicking the link to activate the modal would
disappear when clicked a lot of times very rapidly.

Updated the demo HTML file to show all options.